+++ src/main/java/roomescape/annotaions/Login.java
@@ -1,4 +1,4 @@
-package roomescape.common;
+package roomescape.annotaions;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
