+++ src/main/java/roomescape/auth/JwtTokenProvider.java
@@ -1,4 +1,4 @@
-package roomescape.config.security;
+package roomescape.auth;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.JwtException;
@@ -8,7 +8,6 @@ import java.nio.charset.StandardCharsets;
 import java.util.Date;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-import roomescape.common.Role;
 import roomescape.member.domain.Member;
 import roomescape.member.dto.LoginMember;
