+++ src/main/java/roomescape/auth/Role.java
@@ -1,4 +1,4 @@
-package roomescape.common;
+package roomescape.auth;
 
 import java.util.Arrays;
 
@@ -13,7 +13,6 @@ public enum Role {
                 .findAny()
                 .orElseThrow(() -> new IllegalStateException("일치하는 권한이 없습니다."));
     }
-
 
     public boolean isAdmin() {
         return this.name().equals("ADMIN");
