+++ src/main/java/roomescape/member/dto/LoginMember.java
@@ -1,6 +1,6 @@
 package roomescape.member.dto;
 
-import roomescape.common.Role;
+import roomescape.auth.Role;
 
 public record LoginMember(Long id, Role role, String name, String email) {
 }
