+++ src/test/java/roomescape/exception/DateRangeValidatorTest.java
@@ -1,4 +1,4 @@
-package roomescape.config;
+package roomescape.exception;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
@@ -8,6 +8,7 @@ import java.time.LocalDate;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import roomescape.exception.DateRangeValidator;
 import roomescape.reservation.dto.ReservationSearchCondRequest;
 
 class DateRangeValidatorTest {
